Refactors daily log component
Renames and restructures the daily log component to enhance modularity and reusability.
The changes involve renaming the component for better clarity and modifying its structure to improve its integration within the page.
The post component was removed as it was not used.
Adds prisma formatter to vscode settings.
Fixes relative imports for layout components.